refactor(engine): platform shelf dissolves into the engine — manifest-io and schema to kernel, discovery-utils to domain
workflow-shared/scripts is empty and gone: manifest-io.cjs and
manifest-schema.cjs land in the kernel (the manifest's on-disk contract —
IO discipline and status vocabulary), discovery-utils.cjs lands in the
domain ring (lifecycle joins, next-phase computation, cache status —
derivation logic, not mechanism). Every require re-pointed: engine
internals go sibling-relative, the five per-skill read adapters point into
the engine domain, the manifest CLI consumes the kernel modules until its
own absorption. Source-pin tests updated deliberately; sandbox fixtures no
longer copy a shared shelf that no longer exists.
