+++ skills/workflow-bridge/scripts/discovery.cjs
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const { loadManifest, phaseStatus, fileExists, listFiles, listDirs, computeNextPhase } = require('../../workflow-shared/scripts/discovery-utils.cjs');
+const { loadManifest, phaseStatus, fileExists, listFiles, listDirs, computeNextPhase } = require('../../workflow-engine/scripts/domain/discovery-utils.cjs');
 
 const ALL_PHASES = ['research', 'discussion', 'investigation', 'scoping', 'specification', 'planning', 'implementation', 'review'];
