+++ skills/workflow-engine/scripts/domain/discovery-utils.cjs
