+++ skills/workflow-engine/scripts/domain/epic.cjs
@@ -7,7 +7,7 @@
 // Reads the work unit's manifest (already loaded by the caller) and computes
 // per-phase items, lifecycle joins, gating, next-phase readiness, and the
 // discovery map. Pure over its inputs: same manifest, same answer. Shared
-// manifest semantics come from workflow-shared/discovery-utils — never
+// manifest semantics come from domain/discovery-utils — never
 // duplicated here.
 // ---------------------------------------------------------------------------
 
@@ -21,7 +21,7 @@ const {
   computeAnalysisCacheStatus,
   compareMapRows,
   computeNeedsSequencing,
-} = require('../../../workflow-shared/scripts/discovery-utils.cjs');
+} = require('./discovery-utils.cjs');
 
 const EPIC_PHASES = ['discovery', 'research', 'discussion', 'specification', 'planning', 'implementation', 'review'];
