+++ skills/workflow-engine/scripts/domain/transitions.cjs
@@ -24,7 +24,7 @@ const { loadWorkUnitManifest, saveWorkUnitManifest, withWorkUnitLock } = require
 const { commitScopedWithKb } = require('./commit.cjs');
 const { knowledge, INDEXED_ARTIFACTS } = require('./kb.cjs');
 
-const { VALID_PHASES, VALID_PHASE_STATUSES } = require('../../../workflow-shared/scripts/manifest-schema.cjs');
+const { VALID_PHASES, VALID_PHASE_STATUSES } = require('../kernel/manifest-schema.cjs');
 
 // Phase-item lifecycle operates on WORK phases only. Discovery items are map
 // items (no lifecycle status — computed at render time); they are created and
@@ -32,7 +32,7 @@ const { VALID_PHASES, VALID_PHASE_STATUSES } = require('../../../workflow-shared
 const LIFECYCLE_PHASES = VALID_PHASES.filter((p) => p !== 'discovery');
 
 // Refuse any status write the manifest CLI would refuse — the two enforcers
-// share one schema (workflow-shared/scripts/manifest-schema.cjs), so the
+// share one schema (kernel/manifest-schema.cjs), so the
 // engine can never be the permissive path around a validation refusal.
 /** @param {string} phase @param {string} status */
 function assertLegalWrite(phase, status) {
