+++ skills/workflow-engine/scripts/engine.cjs
@@ -20,7 +20,7 @@ const { signpost, box, wrapWithPrefix, renderTree, WIDTH } = require('./kernel/r
 const { loadWorkUnitManifest, saveWorkUnitManifest, withWorkUnitLock } = require('./kernel/manifest.cjs');
 const { commitScopedWithKb } = require('./domain/commit.cjs');
 const { addSubtopic, setSubtopicState, mapState, SUBTOPIC_STATES } = require('./domain/discussion-map.cjs');
-const { VALID_ROUTINGS } = require('../../workflow-shared/scripts/manifest-schema.cjs');
+const { VALID_ROUTINGS } = require('./kernel/manifest-schema.cjs');
 const { sequenceMap, addItem, editItem, removeItem, renameItem, rerouteItem, handleItem, reactivateItem } = require('./domain/discovery-map.cjs');
 const { startTopic, completeTopic, supersedeTopic, cancelTopic, reactivateTopic } = require('./domain/transitions.cjs');
 const { initTasks, startTask, fixAttempt, completeTask, analysisCycle } = require('./domain/tasks.cjs');
