+++ tests/scripts/test-engine-transactions.cjs
@@ -805,7 +805,7 @@ describe('knowledge store rides along on every engine commit', () => {
 });
 
 describe('schema enforcement: engine refuses what the manifest CLI refuses', () => {
-  const { VALID_PHASE_STATUSES } = require('../../skills/workflow-shared/scripts/manifest-schema.cjs');
+  const { VALID_PHASE_STATUSES } = require('../../skills/workflow-engine/scripts/kernel/manifest-schema.cjs');
 
   it('discovery is not a lifecycle phase — start/complete/cancel all refuse', () => {
     const dir = setupGitFixture();
@@ -829,7 +829,7 @@ describe('schema enforcement: engine refuses what the manifest CLI refuses', () 
     assert.deepStrictEqual(VALID_PHASE_STATUSES.discovery, []);  // map items carry no status — empty vocabulary refuses every write
     const src = fs.readFileSync(
       path.join(__dirname, '../../skills/workflow-engine/scripts/domain/transitions.cjs'), 'utf8');
-    assert.ok(src.includes("require('../../../workflow-shared/scripts/manifest-schema.cjs')"),
+    assert.ok(src.includes("require('../kernel/manifest-schema.cjs')"),
       'transitions must require the shared schema, not mirror it');
     assert.ok(!/VALID_PHASE_STATUSES\s*=\s*{/.test(src), 'no local copy of the status table');
   });
